requirements.txt
langchain
python-dotenv
langchain-openai
langchain-community
beautifulsoup4
ipykernel
faiss-cpu
streamlit
